Several features added
Throw white blood cell
All the bugs can be confidured to be (or not) sensible and how much to
white blood cells, soap drops, antibiotics...
Changed the Microbes script to deal with the projectiles collissions.
(still work to do with the be washed away animation)
Changed the way of setting the bounds of the camera
Changed the way that the player appears in the screen to support
instantiating amy or harry depending on which one is selected.