Added final screen with the winner
Fixed the problem when at the beginning the tv animation was played and
the text should be displayed immediately but it wasn't.
In the introduction of each level, the text displayed over the phone
can be skipped with any key, making it easier to skip.
Full screen disabled (only in the build on Heroku, because the change
to make it must be done to the html file)